Fix issue with the height of tinymce editor.
Improve the bridged-code.
Tinymce init function called while target textarea gets rendered instead of resized.
Remove the cloning of tinymce config as it is not necessary as we instantiate editor separately.
Add a check in Html parser to make sure that argument is not undefined to avoid JS error.
We are registering the resize event on textarea component when htmlEditor gets initialized,
which lead us to miss some of the resize calls.
Logically, we have to wait for text area component before adding any event on the same.

References: KW-81