rules: open another user's mail filters
Add the ability to open another users mail filter, this feature is
limited to only allow opening of already fully opened user stores and
folder permissions on the inbox.

The JavaScript webclient sends the store_entryid of the user along with
every rule request so that the backend knows in which store the rules
should be opened or saved.

Only give option to edit rules of shared stores when user has owner
permissions on the full store.

Added configurability of rules for others users

References: KW-158