Implement hierachy unread counters notifications
Previously a user only received new mail notifications when the Inbox
folder state changed between the start and end of a request.

Every request the inbox folder was checked for updates, adding this
functionality to all subfolders wouldn't be optimal.

Therefore we introduce a new method of checking the unread count of all
folders in the hiearchy by reading the hiearchytable. The state is
serialized to a file and compared on every remindermodule request (every
30 seconds by default).

When the state is different we send a newmailnotification to the client,
which sends an update of the folder's content_count and content_unread.

JavaScript newmail notification checks if unread_count is not equal to
0, so it won't notify for 0 unread emails.

References: KW-119