Introduce more button for mail dialogs
Add a new more button, which can contain more action buttons.
For the mailpreview it contains copy/move and 'edit as new'.
For the mail tab panel it contains print and 'edit as new'.

Edit as new is now avaliable in every folder in the more button and in
the right click context menu.

References: KW-15