js: drop xregexp library
Remove unrequired xregexp library, used to fix compatibility between old
browser versions for the RegExp object. It also added more functionality
to the RegExp and String.match object which was not used in WebApp.

This also removes the behaviour of annotating *bold* as
<strong>*bold*</strong> and _underline_ as <span
style="text-decoration: underline">_underline_</span> since that uses
XRegExp specific syntax and is unwanted.

References: KW-905