js: remove to be implemented or commented code
Remove codeblocks which are commented out and therefore not used. Remove
components which are "hidden: true" since these will not show in Webapp.
Also remove ambigious code "hidden: false" since that is the default in
ExtJS.

The ConsolePlugin, NonePlugin and the components in GridEditors.js are
not used so remove these files.

References: KW-425