KW-237 : Facility to load shared contact folders.
Introducing a config option which decides weather the shared,
contact folders would be displayed in Addressbook or not.
Adding respective default value for the config option.
Conditionally adds available shared contact folders to profile-section,
with every new mapi-session as per the limitation of mapi.
Optimizing the code to execute only required part.
Adding functionality to only display shared-contact-folders which are opened.