Create two newlines when replying plaintext
When using plaintext mode and replying, reply all and forwarding a mail
the Webapp creates four newlines, while in HTML mode we create two
newlines. The plaintextQuotesTemplate never checks if the signatureData
is set, so four newlines are created.

Check if signatureData is set and then add three enters, since the
signature data does not contain an enter. When the signatureData is set
we do not need two enters since the signatureData contains two enters
before the actual signature text.

References: KW-626