js: only register stores when presenceplugins are loaded
When there are no presence plugins enabled do not register stores to
obtain user information. This fixes performance issues when loading the
GAB with a lot of users and when no presence plugin is not enabled.
The performance plugin with a large GAB and a presence plugin enabled is
not resolved.