KW-1200 : Fixing meeting request error
We are synchronizing whole record received in mail notification
to just relay the changes to ShadoStore record.
This is causing some issues as the notification record received
from server doesn't contain all the properties Shadow record required.
To overcome this after effect, we just restrict the synchronization
of MailStore-record with respective ShadowStore-record to message_flags
property only.
message_flags property needs to be synchronized to fix the original case
described/ reported in KW-69.