Hide notification sliders by clicking on them
- Default register the click event for all slider
notifications which are not persistent.
Event handler is used to trigger notification for
discarding existing notification.

- Added stopFx flag in removeMessage function
of SliderContainer.js which used to discard
the notification.

- Refine the code for show notification in
MessageContentPanel.js where removed this.sendingEl
and persistent config from notification config options
as sent and saved  notifications are not persistent notification.

- We dont need to trigger discard notification for sending notification in
hideInfoMask function of MessageContentPanel.js because container of notification
is create content panel which destroy after sending message,
so notification also gets destroyed automatically.

- Refine show notification related code of DelayedDeliveryContentPanel.