fingerprint: make checking fonts twice as fast
Make checking installed fonts twice as fast by doing less. First of
precalculate the width and height of our comparison container for
monospace and arial. This is done when the checkfont object is created.

The main consumer of checkfont is the fingerprint object which checks a
full list of fonts, therefore don't recreate the to be checked container
every time we check for a font.

References: KW-1464