client: cache JavaScript load order on LOAD_SOURCE
When loading WebApp in a development env. the JavaScript loader order is
calculated every time the WebApp is loaded. Caching this order has a
serious performance benefit and leads to a 94% decrease in loading time.

Caching is done by writing the <script> loader order to a file and
calculating an md5 of all the JavaScript files in client/ and plugins/
if this md5 changes a new load order will be written to the file.

The cache file unique per WebApp location on a developers system.

References: KW-1370