KW-1098 : Facility to choose calendar for new MR.
Adding a drop-down list for all the opened calendars to choose into
which calendar the new appointment or meeting-request will be created.
The drop-down list consist of all the opened calendars irrespective of
the necessary modification rights.
If current user doesn't have necessary rights to do so, an exception
will be raised while saving particular item.