mail: implement lazy loading of unread email items
References: KW-787

After the loading of the mail list the unread mail items will be fetched
in batches (configurable by the admin)

MAPIStore: support opening items in batch

Allow Store.open() to open multiple records in the same request,
previously only one record could be opened at once. This change is
required for the lazyloadmail plugin which preloads unread emails in the
mail context for performance.