KW-2028 Attachment should send when Address book is open while auto-saving
- Added PR_EC_KW_ATTACHMENT_ID
custom property for the attachment records.

- It should be treated as a quine id
of attachment. We need this property because
of merging sub stores when autosaving mail.

- Changed attach_num  as an id property to attach_id (PR_EC_KW_ATTACHMENT_ID).

- earlier attach_num is used as an id property for
the attachment record which creates a problem because
attach_num for unsaved attachments is -1 which is ID
property so all attachments have the same id and because
of that while saving attachments in the record, sub store use
Add function of an ExtJS which only save one attachment
and skipped rest.

References : KW-2028