KD-310 : Fixing zoom issue in DeskApp
Issue is being reproduced only while the message body is focused.
The root cause of the issue, where zoom functionality isn't working
properly while using "CTRL + Wheel", is that the "wheel" event from
and iframe is not relayed to the main window.
Adding facility to relay the "wheel" event from iframe to main
window object.