Add KC_FILTERED_BODY configuration option
The new KC_FILTERED_BODY option makes use of the new property
PR_EC_BODY_FILTERED which is the filtered on the server instead in
WebApp. Therefore skipping the costly HTML filtering in PHP, this can
potentionally speed up loading HTML bodies in WebApp. This feature is
available from Kopano Core 8.5.0, for older versions a fallback
mechanism is created and the PR_HTML property is loaded.

References: KW-2077