KW-614 : Adding functionality to import vcf
Adjust import functionality, introduced in KW-612, to allow user to
import attached vcf file into selected folder.
Filtering the folder-tree in import-dialog to show only contact
folders.
Enable import-context-menu only if the functionality has support at
server side, adding a conditional-tooltip to inform user otherwise.
Check the availability of mapi function to check
if the functionality is supported or not.