KW-2343: Check if tab has a dialog before using it
The ZDeveloper plugin creates a tab to show the insertionpoint but does
not create a dialog for the tab. This broke the WebApp because the
functionality that checks if a tab is already open assumed that there
was always a dialog and only checked if the dialog had a record
property.