KW-2261 : Fixing bad value in create-in combo.
Issue is when user doesn't have enough rights for the selected
folder then it is not available in combo-store.
And as a selected folder, combo box gets the initial value of
that particular folder only.
As a solution we are now selecting default folder when user
don't have enough rights for selected folder.