Adds check if user can get to the inbox of a shared folder
The Rules panel in settings creates a dropdown with users for whom the
current user can create rules. When this dropdown is populated we check
if the user has folder permissions on the other users inbox. However we
forgot to check if the user can get to inbox of the other user at all.
If the user did not have this permission the settings context broke.
This commit adds that check.