Make sure user_name prop of MAPIStoreRecord is a string
Ref: KW-2474

If we don't explicitly declare the user_name property as a string, Ext
will try to find it's type. This means that a user_name that is a number
will be returned as a number instead of string by the store.get()
method and comparisions can fail. e.g:

if ( store.get('user_name') === userName )

is false if userName is '1234' because 1234 !== '1234'