KW-2575 : Fixing canceled attachment upload issue
I know that what is implemented here is not proper as far as the
architecture and code quality are concerned.
But this is the only feasible way to abort the upload.
We simply want attachID on client side to make delete request, but
attachID is not available on client side until the upload process gets
finished and response arrives back.
We can't even send attachID with respective File object, as File object
is read-only.
As a solution, i just create a clone of File and append unique-id
to file name.
Handling embedded attachment case as well.
Handling aborting attachment as File API isn't supported by IE/Edge.
Disabling attachment removal when it is uploading because we can't provide
attachID from client-end to identify uploading attachment with IE/Edge.