KW-2587 : Disabling close button while uploading
Close button will remove the attachment while uploading are still
going on. this is something needs to be prevented.
Check before process the removal of attachment if uploading is still
going on or not.
Don't remove the attachment if uploading.
even while pressing delete and backspace buttons from attachment field.