Don't let closure compiler add polyfills
Ref: KW-2446

The closure compiler by default adds polyfills that are needed to a
global object called $jscomp. Every compilation adds its own $jscomp
object thereby overwriting existing objects. (a bug is fixed with
this commit:
https://github.com/google/closure-compiler/commit/f17feaad720aa06b3cfe1e44045add872e2e1a50)

So we could fix it by updating the compiler, but this means that every
plugin must be updated to a version that is compiled with the newer
compiler. So instead we will tell the compiler not to add its own
poyfills. WebApp already includes the polyfills it needs itself, so
this isn't abundant anyway.