Remove reminderlistmodule checksum
Remove the server hashing of the list which is send to the client, since
the client does not use this hash to verify if the data has changed. The
client verifies if the data is changed by caching a JSON object in JS
and verifying if that has changed.

References: KW-2650