Add unread counter updating for shared stores
Implement a poller in the webclient which by default every 15 minutes
queries the backend for a sharestoreupdate. The hierarchymodule acepts
the request and calls the notifier for every opened shared store that
qualifies for an update (these are inbox and all folders). The hierarchy
update mechanism for a logged in user is extended to handle shared
stores as well allowing more re-use of code, the code logic is the same
query the hierachytable and store this in the session data. Compare it
on the next request against the saved data, if the data is changed
inform the webclient.

References: KW-361