mapisession: add getServerVersion
Implement retrieving the server version from the actual Server. This is
done by retrieving the PR_EC_SERVER_VERSION from the logged in users
store. Since this is a new property check if it exists and otherwise
return an empty string and fallback on the previous default behaviour.

References: KW-1718