server: properly check if folders are accessible
The previous check for folder permissions, only checked the second
folder returned by getHierarchyList. The new code to fetch the
folderlist is quicker but returns the results different order. Change
the code to count the number of accessible folders and compare them to
the the total amount of folders. Since a full store might be partly
accessible.