server: Generate entryid for recipients without entryid
When a PST imported and users can't be resolved since they where removed
for example, the user has no entryid. This causes issues when double
clicking, opening the item in a tab which makes the JavaScript code
mergge substores which uses an entryid for comparison.

The solution is to on the fly generate an oneoff entryid for the
recipient, this has the downside of not saving the entryidtherefore the
issue will exist indefinitely.