Make check if SSO request more strict
Some environments might always sent $_SERVER headers even when not in
use (then the value is empty). WebApp would think that it is a SSO
request when the REMOTE_USER variable is set even when empty. This
change adds an additional condition that only if the REMOTE_USER is set
and not empy the request is treated as a SSO request. This makes WebApp
username and password login work when used with Kweb.

Until this is merged, the a workaround is to set
DISABLE_REMOTE_USER_LOGIN to true in WebApp config.php.

Fixes: KW-2922