Set default font without being too restrictive
Ref: KW-2979

This commit chnages the behaviour of the CSS of the WebApp that uses the
wildcard rule to set the default font on all elements. Now we will set
the default font on the body and on all elements that have been
overridden by the CSS of ExtJS. This commit also removes the !important
directive of the default font.