Fix build
Ref: orphan

The build tried to create the app-icons scss file because the client
target depended on all the scss files. (they are needed for the source
maps)
This commit removes the dependency of the client target on the scss
files. Instead it will just bluntly copy the scss files to the build
destination.