Implement Conversation View
Ref: KW-3160

This commit adds Conversation View mode to the WebApp. A setting is
added to make it possible for the user to enable the Conversation View.

Conversations are built up from items from the users Inbox and Sent
Items folders. Items are grouped based on normalized subject.

A search folder is created with all items from the Inbox and the Sent
Items folder. From this searchfolder items are fetched in batches of
1500 items. The items in these batches are then grouped into
conversations.

The conversations will rendered in the mail list as a 'header item'
with the real items collapsed under it. To do this the backend will add
these fake header items to the list it has generated.