Load only enabled plugins
Ref: KW-907

Only php files of enabled plugins will be included and instantiated.
Only js and css files of enabled plugins will be sent to the browser.

To be able to know which plugins are enabled by default by the admin a
new define has been introduced in the config.php of the WebApp:
DEFAULT_ENABLED_PLUGINS_LIST
The 'enable' setting in the config.php files of plugins will no longer
be processed because we must know which plugins are enabled before we
will load them (and they can inject their settings).

The meta data of plugins will be fetched from their manifest file. The
title element will be used as name. The 'allowuserdisable' and
'allowuservisible' properties will now be taken from the manifest and
not from the values that are set in the javascript when the plugin
registers itself.

The default 'enable' settings are no longer set in the plugins
config.php files.
The display name is no longer set when registering the plugin in the
JavaScript.
The 'allowuserdisable' and 'allowuservisible' have been moved to the
manifest.
The 'name' element has been removed from the manifests. (it is now taken
from the plugin directory name)

Catch Exception thrown by invalid XML in manifest