Fix 500 error when CLIENT_TIMEOUT is set
When the CLIENT_TIMEOUT is set invalid JSON from the fingerprint and
OIDC token request a JSONException is raised which is not compatible
with the build-in JSONException since it has no getDisplayMessage.

Instead of trying to parse invalid JSON check if it was a JSON request
and otherwise return.

References: KW-3398