language: add full en_GB translation
To get British date and time formatting for sure, a .po file is a _must_.

Adding en_GB to the Settings dialog alone does not magically make
grommunio-web British; what it actually does is cause the UI to go into
*Untranslated* mode and use whatever the internal components (e.g.
extjs) use for untranslated strings. That _may_ be British, but it also
may be not.