smime: avoid saving drafts with bogus SMIME PR_MESSAGE_CLASS
When using the Sign/Encrypt buttons, the smime plugins changes
PR_MESSAGE_CLASS from IPM.Note to IPM.Note.SMIME*, but does not
transform the rest of the message to conform to the OXOSMIME
specification. This causes drafts to be unprocessable by Outlook and
Gromox. Only when the draft is sent (submitted) will the smime plugin
perform the OXOSMIME transformation, at which point OL and Gromox can
read it again.

Switch to a different message class to indicate "to be
SMIME-transformed later" which does not have a special meaning.

References: vcs/web/216