fix: insecure temporary file handling for thumbnail generation
When an user saves his preferences in the grommunio web interface
a avatar thumbnail is generated and written to a temporary file.
The filename is derived from a static prefix, the current unix epoch
timestamp and the image type. This corresponds to the common attack
vector [CAPEC-149](https://capec.mitre.org/data/definitions/149.html)
*Search for predictable names for temporary files*.

Without a specific intention, a race condition can occur between
users who triggers a preview creation at the same time with a
granularity of one second. This can lead to access corrupted or
unexpected image data.

The temporary file is created in `/tmp` with world readable
permissions. This corresponds to the common attack vector
[CAPEC-155](https://capec.mitre.org/data/definitions/155.html)
*Screen Temporary Files for Sensitive Information*. This is considered
to have a low risk as it concerns only contact avatars, even if they
are not available to the general public.