fix: clear translation cache if no translation data was found
We observed missing translations after upgrades/reboots. It seems that
the content of the lookup variable (`($memid, 0)`) was an empty array.
Grommunio-web did not recover from this situation as it never clears the
cache if the lookup table is empty, automatically. A workaround is to
clear the cache using `ipcrm -M 0x950412de`, manually

This change clears the cache when the lookup variable is empty or the
translation is not found. This allows to recover from a broken cache,
automatically.