fix: unable to resolve distlist in sent message
The email address of an AB distlist was its SMTP address, but its
type was EX which caused that the email address was not saved
correctly in the recipient table.
Also fix warnings that grommunio-web is not able to open a store
when checking details of such a recipient.

References: Issue #199