fix: clear translation cache if no translation data was found
We observed missing translations after upgrades/reboots. It seems that
the content of the lookup variable (`($memid, 0)`) was an empty array.
Grommunio-web did not recover from this situation as it never clears the
cache if the lookup table is empty, automatically.

While the cache is being normally cleared by grommunio provided packages
via ipcrm -M 0x950412de, this change does not require this additional
step.

This change clears the cache when the lookup variable is empty or the
translation is not found. This allows to recover from a broken cache,
automatically.