Display folder rights for granted users
The user will only be able to see rights of the shared stores and
folders only if he has owner permissions. Otherwise the user will
only be able to see his own permissions if set as well as default
and anonymous permissions.

The user will be able to see the details store/folder information
like size and item count only if he has owner permissions.

References #356