server: stop throwing No Access error when opening read-only embedded messages
g-web invokes openMessage with the parse_smime set to true, but
``parse_smime`` has the side effect of wanting to modify the message
on the assumption that ``$message`` is a nascent (unsaved) temporary
object. As a result, the mapi_setprops call inside parse_smime fails
and throws an error:

.. code-block:: php

        // mark the message as read if the main message has read flag
        if ($read) {
                $mprops = mapi_getprops($message, [PR_MESSAGE_FLAGS]);
                mapi_setprops($message, [PR_MESSAGE_FLAGS => $mprops[PR_MESSAGE_FLAGS] | MSGFLAG_READ]);
        }

The fix is to not invoke SMIME logic for non-SMIME messages.

References: vcs/gromox/646